Added DisabledSchemes policy.
This is a list of strings. Any protocol scheme listed in this policy is
disabled, and URLs using that scheme won't load (either from the Omnibar, links,
bookmarks, or when requested from the renderer to the browser).

Also introduced ListPrefMember, to track a ListValue preference.

"Virtual" or "pseudo" protocol schemes such as "about" and "view-source" are
also handled.

BUG=57477
TEST=unit_tests, use a policy to disable specific schemes (e.g. "file", "ftp")

Review URL: http://codereview.chromium.org/6712065

git-svn-id: http://src.chromium.org/svn/trunk/src@82013 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: efeb985df2cf18dff7ddd4066b3dcb85c6205da9