Change the HTTP cache to cache the entire certificate chain for SSL sites
When persisting an X509Certificate to a pickle, such as when storing to the HTTP cache, persist any intermediate certificates in addition to the end-entity certificate. This will allow the complete certificate chain to be displayed to the end user when viewing a cached entry, independent of whether a network request has been made to that site during the browsing session.

R=agl
BUG=7065
TEST=X509CertificateTest.Persist

Review URL: http://codereview.chromium.org/4645001

git-svn-id: http://src.chromium.org/svn/trunk/src@82214 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fd3272301b63b3fe6d4b1fc0b45624c59c6f84e2