Scroll: enable scroll viewport for find bar and anchor line
When doing finding, browser can't scroll the correct visible area
for a searched text. That's because the backend always transfers
incorrect scroll position to me. Browser now set our viewport
according to render process.

Current implementation has one issue to be improved: we always set
viewport as the scroll position from render process. The reason is
we have to do it when doing an internal page jump(#seg). It's not very
friendly to user to doing finding. If a next/previous string is


Former-commit-id: 6156652f9a8743947e917c7815106136fb5689f8