Merge 83936 - Have the phishing classifier guard against extra PageCaptured calls.
Currently, the classifier assumes that any PageCaptured call will be preceded by a DidCommitProvisionalLoad.  This seems to not always be the case, and is causing occasional crashes.  Cancel classification in this situation so that we don't swap out the page text from under the term feature extractor.

BUG=80325
TEST=PhishingClassifierDelegateTest.DuplicatePageCapture

Review URL: http://codereview.chromium.org/6902171

TBR=bryner@chromium.org
Review URL: http://codereview.chromium.org/6935012

git-svn-id: http://src.chromium.org/svn/branches/742/src@84161 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6b06cce1fbaee8305f8657e0cca930f8a88848aa