ouble tap: Using scale to improve double tap experience
Scale web page without re-layout. We calculate the best scale
factor and scale center point from screen size and the element
area that user double tap at. The Qt widget used to display
page content is scaled to the best scale factor with scaling
animation. The animation of double tap is the same as the pinch
gesture now.

The side effect of this approach is, pure text HTML page which
fits exactly the whole screen will not be zoomed in. As a result,
double tap has no effect for this kind of page.


Former-commit-id: a216b6ae56b0a64724c459d58a6c4fa9aa98821e