UX: tab sidebar is not align with visual design
The current design is not accroding with disual design, including
borderimage, font color, font size, alignment, with, height and so on.

In addition with the improvement of UI, a new feature is added in this
commition, the warning message will be displayed, when the tabs number
exceed the limits.

There is a limitation in this current version, the blank thumbnail can't be
added in the tab sidebar, when you taps the new tab item.


Former-commit-id: f806928dc7d8d352a443502d5462ac856287c347