Merge 85563 - Initialize all values in BugReprotData.
On ToT I found that if I issued a bug report quickly it would crash because the syslogs data was not ready, presumably because fetching that data was put on a separate thread.
I'm trying to confirm that this is the same issue as the cras repport in chromium-os:15253, but regardless this simple fix should be in the M12 branch.

BUG=chromium-os:15253
TEST=Report an issue immediately after logging in. Chrome should not crash.

Review URL: http://codereview.chromium.org/7016045

TBR=stevenjb@google.com
Review URL: http://codereview.chromium.org/7030030

git-svn-id: http://src.chromium.org/svn/branches/742/src@85571 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1163d17899dc7dced2e0defdc1c8b240aca9f1ef