Fix bug #17746: cannot show more tweets of twitter.com and weibo.com when scroll to page bottom
twitter.com and weibo.com uses window's scroll events and scrollTop etc.,
to decide if show more tweets.
The root cause is when delegating the scrolling to front-end, the scroll
event fails to trigger and the scrollTop is not set correctly.

Performance:
To trigger scrolling events will incease traffic of IPC. In touch FPS test,
the current implementation gots same FPS count as previous one. So commit it.

Tests:
1. twitter.com and twitter.com
2. http://demos.9lessons.info/load_data_scroll.php

Former-commit-id: 8655227a7dc226ded753004a60d04a0a5fad5540