Browser service: optimize the timing for capturing thumbnail
In case of the following situation, a thumbnail for a web page
will be captured:
1. a tab is inserted (TabInsertedAt)
2. a tab is changed if and only if the tab is the selected one
   and it finishes to load (TabChangedAt). Only care about the
   selected tab and ignore other hidden tabs to improve browser
   performance.
3. a tab is selected.
4. a tab is deselected (TabDeselected)

In addition, the time interval for capturing thumbnail is set
to 5s to avoid frequent operation of catpuring thumnail.

TODO:
Need to reduce thumbnail capturing further if a page is already
loaded successfully. No need to capture thumbnail again if a
page has been captured.

Former-commit-id: 5001bbbabcda7cf0b81b44eaa6a307969c2968ef