Interstitial: don't set render_host as main host for interstital page
This can cause a crash. To reproduce it, you can follow up:
1) set home page as https://meego.com
2) make sure your date is invalid, such as 19600101
3) start browser then you can see ssl error
4) open a new url in current tab, such as google.com
5) after loading is completed, click reload button

The root cause is that browser creates 2 render view host, one is
for main content page, anther one is for interstitial page to show
ssl error message. Current implementation sets its rwhv_ and
rwhv_view_ each time a new render view host and rwhv widget is created.
However, for interstitial, it should not be set as the main content
render. When switching to a new page, the interstitial will be destroyed
but we still use their pointers in tab contents view. When opening a new
page, the render view host for main content page is reused. If clicking
reload button, the tab contents will use the pointers to set focus.
This causes the crash.

One following improvement is to create a new render view host and destroy
the render host for meego.com though its render host for meego.com is not
used. This can avoid


Former-commit-id: 869343979db4938759d359cd9f356d9d69576274