UX: show window a bit later to avoid hole in status bar
To save startup time, we show main window ahead of time.
But when startup, the left corner of status bar will have
a hole. This is not very friendly to users.

To avoid this issue, don't show main window until the qml
UI is loaded. But this may hurt startup time by increasing
about ~50ms.

Reviewed by Ningxin.


Former-commit-id: 22fb9f6b65d82e1d05e4d98d71cfce8543e9876b