UX + Tiling: avoid showing checkers and many page jumps
Currently when starting a new page, reloading a page or orientation
is changed, there are many page content jumps and checkers will
be painted many times. This is not user friendly.

For many checkers are shown on loading a new page because the contents
is changing many time during this period, there are 2 problems to
be resolved:
1) accept the first pixmap from backend to save performance
2) keep them and send the new tile requests when content size is
changed

The solution is to hold the current pixmap and send paint new rects as
well. This can still paint old pixmaps while waiting for new updates.
So if the tile is change during the content size change, we update
it when needed. But this causes a problem that when orientation is
changed, some tiles can't be repainted. The solution is that keeping
those tiles whose rectange is the same its pixmap when the contents
width is not changed and height is changed.

Reviewed by Hongbo and Ningxin.


Former-commit-id: 646aa55c2fd83e7caaa1aecf99faf6badd2faa8d