UX: Add crash tab dialog into dialog queue
When the tab page is crashed, crash tab dialog will be popped up. But if
there are server dialogs need to be popped up, the first dialog can be
displayed, and others will be discarded.

This commit has fixed the bug, the way is that the dialog was added
into dialog queue, when it was triggered. And always the first dialog was
displayed, the next dialog was popped up, only after the front dialog
exited.

Reviewed by Jamse[james.wei@intel.com]


Former-commit-id: f899a5cb06b56df0d9fb9fdf8ef9e7bb14ef8805