UX: Add ssl dialog into dialog queue
SSL dialog is s kind of popup dialogs which include js dialog, crash
dialog and others. We use an dialog queue to control these dialogs so
they could pop up in a queue order. When a dialog is triggered, it`s
added into dialog queue, and always the first dialog is displayed, the
next dialog was popped up only after the front dialog exited.

reviewed by Yin li


Former-commit-id: 86fbfb72ba2762205a85b88da439f02cd3922ff7