Omnibox: show correct text when focus in and page is loading [BMC#20080]
Omnibox shows incorrect text if user focuses in omnibox when page is
loading. Reproduce steps:
1. open a website (the more complex the better)
2. click into the omnibox
3. select the first suggestion item
4. click omnibox again as quickly as you can
5. omnibox shows page title other than url

That's because in step 3, browser will load the selected item in popup
list. After the page is loaded, page title is set. To solve this, add a
flag to indicate if page title is got for tabcontent. If yes, omnibox
will not update urltext in focus in mode.

Reviewed by James.


Former-commit-id: 2003e158c465565aefcb0dd41515f86333930384