BMC#19270 - Scroll up/down on web page while VKB is open will get unwanted page shift
Because mail.google.com is iframe, and only accept mousewheel event
achieve pan effect, not as other page by move flickable.

Right now, when vkb is popup, it can pass event to browser, so browser
can pan the page(flickable) with vkb is showing. So to get the same
effect on iframe page, not to stick inputtext at the same place.

Review by deqing.


Former-commit-id: 7584cc4f23eac4db35bb5856564f2e05acbb424e