Fix BMC 21611 image for missing plugin is cropped when plugin is narrow
In the feature of click to play, if the window size of displaying plugin is
smaller than plugin icon, the plugin icon can't be shown fully, so we
must change the size of plugin icon according with the window size of
plugin.

And if the window size is too small. the plugin icon will not be shown.
Test cases:
1. www.sina.com.cn
2. www.youku.com
3. www.sohu.com

Reviewed by James[james.wei@intel.com]


Former-commit-id: 4c2b6f75a1b2c2f02bdb5cf181825e0e3716e638