Fix BMC 21924 [REG] Bookmark Manager fails to change its layout according to orientation
This bug was introduced by commit d526d3d88 of Session restore to enable
bookmark manager restore.

The restore code added a new way to show bookmark manager, but it sets
portrait property to a fix value, which makes bookmark manager fails to
change its layout according to system's orientation on the fly.

This commit also removed out-of-date way of showing bookmark manager.

Reviewed by Gao Chun.


Former-commit-id: b1840b03ce17f24ed358560bb24caf3e2867df3a